requirements.txt
# Pin Torch CPU wheel and related libs
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu

# Core app
Flask==3.0.3
gunicorn==21.2.0
ultralytics==8.2.79
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
